+++ test/node_smoke_tests/commonjs/regular/window_present_originally.cjs
@@ -2,9 +2,9 @@
 
 const { JSDOM } = require( "jsdom" );
 
-const { ensureJQuery } = require( "./lib/ensure_jquery.cjs" );
-const { ensureGlobalNotCreated } = require( "./lib/ensure_global_not_created.cjs" );
-const { getJQueryModuleSpecifier } = require( "./lib/jquery-module-specifier.cjs" );
+const { ensureJQuery } = require( "../lib/ensure_jquery.cjs" );
+const { ensureGlobalNotCreated } = require( "../lib/ensure_global_not_created.cjs" );
+const { getJQueryModuleSpecifier } = require( "../lib/jquery-module-specifier.cjs" );
 
 const jQueryModuleSpecifier = getJQueryModuleSpecifier();
