+++ test/node_smoke_tests/module/factory/document_missing.js
@@ -1,0 +1,13 @@
+import assert from "node:assert";
+
+import { ensureGlobalNotCreated } from "../lib/ensure_global_not_created.js";
+import { getJQueryModuleSpecifier } from "../lib/jquery-module-specifier.js";
+
+const jQueryModuleSpecifier = getJQueryModuleSpecifier();
+const { jQueryFactory } = await import( jQueryModuleSpecifier );
+
+assert.throws( () => {
+	jQueryFactory( {} );
+}, /jQuery requires a window with a document/ );
+
+ensureGlobalNotCreated();
