+++ test/node_smoke_tests/module/factory/iterable_with_native_symbol.js
@@ -1,7 +1,7 @@
 import process from "node:process";
 
-import { ensureIterability } from "./lib/ensure_iterability_es6.js";
-import { getJQueryModuleSpecifier } from "./lib/jquery-module-specifier.js";
+import { ensureIterability } from "../lib/ensure_iterability_es6.js";
+import { getJQueryModuleSpecifier } from "../lib/jquery-module-specifier.js";
 
 if ( typeof Symbol === "undefined" ) {
 	console.log( "Symbols not supported, skipping the test..." );
