+++ test/node_smoke_tests/module/regular/window_present_originally.js
@@ -1,8 +1,8 @@
 import { JSDOM } from "jsdom";
 
-import { ensureJQuery } from "./lib/ensure_jquery.js";
-import { ensureGlobalNotCreated } from "./lib/ensure_global_not_created.js";
-import { getJQueryModuleSpecifier } from "./lib/jquery-module-specifier.js";
+import { ensureJQuery } from "../lib/ensure_jquery.js";
+import { ensureGlobalNotCreated } from "../lib/ensure_global_not_created.js";
+import { getJQueryModuleSpecifier } from "../lib/jquery-module-specifier.js";
 
 const jQueryModuleSpecifier = getJQueryModuleSpecifier();
